Move GFM spec out of tables ext, test strikethrough as well
All tests passing already (there's not many).
