+++ commonmark-ext-gfm-strikethrough/src/test/java/org/commonmark/ext/gfm/strikethrough/StrikethroughSpecTest.java
@@ -1,4 +1,4 @@
-package org.commonmark.ext.gfm.tables;
+package org.commonmark.ext.gfm.strikethrough;
 
 import org.commonmark.Extension;
 import org.commonmark.parser.Parser;
@@ -12,27 +12,26 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
 @RunWith(Parameterized.class)
-public class TablesSpecTest extends RenderingTestCase {
+public class StrikethroughSpecTest extends RenderingTestCase {
 
-    private static final Set<Extension> EXTENSIONS = Collections.singleton(TablesExtension.create());
+    private static final Set<Extension> EXTENSIONS = Collections.singleton(StrikethroughExtension.create());
     private static final Parser PARSER = Parser.builder().extensions(EXTENSIONS).build();
     private static final HtmlRenderer RENDERER = HtmlRenderer.builder().extensions(EXTENSIONS).build();
 
     private final Example example;
 
-    public TablesSpecTest(Example example) {
+    public StrikethroughSpecTest(Example example) {
         this.example = example;
     }
 
     @Parameters(name = "{0}")
     public static List<Object[]> data() {
-        return ExampleReader.readExampleObjects(TestResources.getGfmSpec(), "table");
+        return ExampleReader.readExampleObjects(TestResources.getGfmSpec(), "strikethrough");
     }
 
     @Test
